feat(preview): refresh iframe when strapi updates
